refactor: Uploader demo case
