+++ packages/veui/src/components/Uploader/Uploader.vue
@@ -49,22 +49,14 @@
 </template>
 
 <script>
-import { parse as parseByte } from 'bytes'
 import {
-  keys,
   some,
-  compact,
-  last,
   pick,
   includes,
   isArray,
   isString,
   startsWith,
-  endsWith,
-  find,
   findIndex,
-  uniqueId,
-  entries,
   identity,
   omit,
   isNil
@@ -80,6 +72,16 @@ import warn from '../../utils/warn'
 import Lightbox from '../Lightbox'
 import FileUploader from './_FileUploader'
 import MediaUploader from './_MediaUploader'
+import {
+  UPLOAD_UPLOADING,
+  UPLOAD_SUCCESS,
+  UPLOAD_FAILURE,
+  ERRORS,
+  addOnceEventListener,
+  getFileMediaType,
+  UploaderFile,
+  getUploadRequest
+} from './_helper'
 
 config.defaults({
   'uploader.requestMode': 'xhr',
@@ -109,20 +111,6 @@ config.defaults({
   }
 })
 
-const UPLOAD_PENDING = 'pending'
-const UPLOAD_UPLOADING = 'uploading'
-const UPLOAD_SUCCESS = 'success'
-const UPLOAD_FAILURE = 'failure'
-
-const PUBLIC_FILE_PROPS = ['name', 'type', 'poster', 'src']
-
-const ERRORS = {
-  TYPE_INVALID: 'type',
-  SIZE_INVALID: 'size',
-  TOO_MANY_FILES: 'count',
-  CUSTOM_INVALID: 'custom'
-}
-
 export default {
   errors: ERRORS,
   name: 'veui-uploader',
@@ -141,7 +129,7 @@ export default {
       default: 'file'
     },
     value: {
-      type: [Array, String, Object]
+      type: [Array, Object]
     },
     type: {
       type: String,
@@ -211,7 +199,10 @@ export default {
     validator: {
       type: Function
     },
-    maxCount: Number,
+    maxCount: {
+      type: Number,
+      default: Infinity
+    },
     maxSize: [Number, String],
     payload: Object,
     autoupload: {
@@ -260,10 +251,7 @@ export default {
       return this.realDisabled || this.realReadonly
     },
     canNotAddImage () {
-      return (
-        this.realUneditable ||
-        (this.maxCount && this.fileList.length >= this.maxCount)
-      )
+      return this.realUneditable || this.fileList.length >= this.maxCount
     },
     status () {
       if (!this.fileList.length) {
@@ -353,6 +341,9 @@ export default {
     },
     childOptions () {
       return pick(this, sharedProps)
+    },
+    preferType () {
+      return includes(['image', 'video'], this.type) ? this.type : undefined
     }
   },
   watch: {
@@ -384,6 +375,8 @@ export default {
           [[], []]
         )
 
+        const guessType = file => this.preferType || getFileMediaType(file)
+
         // 剩余的按顺序匹配
         let j = 0
         this.fileList = this.fileList
@@ -392,7 +385,11 @@ export default {
               return file
             }
             if (j < rest.length) {
-              file.value = rest[j++]
+              let val = rest[j++]
+              file.value = val
+              if (!val.type && guessType(val) !== file.type) {
+                file.type = guessType(val)
+              }
               return file
             }
             // 处理外部直接减少文件的情形
@@ -403,10 +400,7 @@ export default {
             rest.slice(j).map(val => {
               let file = UploaderFile.fromValue(val)
               if (!file.type) {
-                // 是否应该在 File 内部处理这个逻辑？
-                file.type = includes(['image', 'video'], this.type)
-                  ? this.type
-                  : getFileMediaType(file)
+                file.type = guessType(file)
               }
               return file
             })
@@ -475,7 +469,8 @@ export default {
     },
 
     handleImageAdd () {
-      this.pickFiles(true).then(files => {
+      let restCount = this.maxCount - this.fileList.length
+      this.pickFiles(restCount > 1).then(files => {
         if (!files.length) {
           return
         }
@@ -483,7 +478,9 @@ export default {
       })
     },
     handleImageRemove (index) {
-      this.removeFile(index)
+      // TBC: 是否需要 恢复 被替换的
+      let file = this.fileList[index]
+      this.removeFile(index, { restore: file.isFailure })
     },
     handleImageReplace (index) {
       // TODO: pickFiles 异步回调后可能 fileList index 已经变了
@@ -552,7 +549,9 @@ export default {
                   file.value,
                   this.fileList.indexOf(file),
                   evt
-                )
+                ),
+              oncancel: () =>
+                this.removeFile(this.fileList.indexOf(file), { restore: true })
             })
           })
         })
@@ -574,16 +573,25 @@ export default {
         })
     },
     replaceFile (index, file) {
-      this.fileList.splice(index, 1, new UploaderFile(file))
-      this.uploadFile(this.fileList[index])
+      let newFile = new UploaderFile(file)
+      newFile._replacing = this.fileList.splice(index, 1, newFile)[0]
+      this.uploadFile(newFile)
     },
-    removeFile (index) {
-      let [removed] = this.fileList.splice(index, 1)
-      if (removed) {
-        removed.cancel()
+    removeFile (index, { restore = false } = {}) {
+      if (index < 0 || index >= this.fileList.length) {
+        return
       }
-      this.$emit('remove', removed.value, index)
-      if (removed.isSuccess) {
+
+      let file = this.fileList[index]
+      if (restore && file._replacing) {
+        this.fileList.splice(index, 1, file._replacing)
+      } else {
+        this.fileList.splice(index, 1)
+      }
+
+      file.cancel()
+      this.$emit('remove', file.value, index)
+      if (file.isSuccess || file._replacing) {
         this.triggerChangeEvent()
       }
     },
@@ -601,480 +609,12 @@ export default {
         ...pick(file, ['status', 'loaded', 'total']),
         index
       }
+    },
+
+    // legacy api
+    clickInput () {
+      this.handleImageAdd()
     }
-  }
-}
-
-function addOnceEventListener (el, evt, listener) {
-  el.addEventListener(evt, function callback (...args) {
-    el.removeEventListener(evt, callback)
-    listener(...args)
-  })
-}
-
-function getFileMediaType (file) {
-  const mediaExtensions = config.get('uploader.mediaExtensions')
-  return find(keys(mediaExtensions), function (type) {
-    return some(mediaExtensions[type], ext => endsWith(file.name, `.${ext}`))
-  })
-}
-
-function getMediaExtensionsByType (type) {
-  const mediaExtensions = config.get('uploader.mediaExtensions')
-  switch (type) {
-    case 'image':
-    case 'video':
-      return mediaExtensions[type]
-    case 'media':
-      return mediaExtensions.image.concat(mediaExtensions.video)
-  }
-  return []
-}
-
-function getValidateFileType ({ accept, extensions, type }) {
-  const mediaExtensions = getMediaExtensionsByType(type)
-
-  return function validateFileType (file) {
-    if (!accept) {
-      return true
-    }
-
-    let ext = last(file.name.split('.')).toLowerCase()
-    if (extensions) {
-      return includes(extensions, ext)
-    }
-
-    return accept.split(/,\s*/).some(item => {
-      let acceptExtention = last(item.split(/[./]/)).toLowerCase()
-      if (
-        acceptExtention === ext ||
-        // 对于类似'application/msword'这样的mimetype与扩展名对不上的情形跳过校验
-        (acceptExtention !== '*' && includes(item, '/'))
-      ) {
-        return true
-      }
-
-      return (
-        acceptExtention === '*' &&
-        includes(item, '/') &&
-        (includes(mediaExtensions, ext) || !mediaExtensions.length)
-      )
-    })
-  }
-}
-
-function getValidateFileSize ({ maxSize }) {
-  maxSize = maxSize && parseByte(maxSize)
-  return function validateFileSize (file) {
-    return !maxSize || file.size <= maxSize
-  }
-}
-
-function getCustomValidate ({ validator }) {
-  return function customValidate (file) {
-    return validator ? validator(file) : { valid: true }
-  }
-}
-
-function getValidateFile (options, ctx) {
-  const validators = [
-    [
-      getValidateFileType(options),
-      ERRORS.TYPE_INVALID,
-      file => file.name,
-      identity,
-      () => ctx.t('fileTypeInvalid')
-    ],
-    [
-      getValidateFileSize(options),
-      ERRORS.SIZE_INVALID,
-      file => file.size,
-      identity,
-      () => ctx.t('fileSizeInvalid')
-    ],
-    [
-      getCustomValidate(options),
-      ERRORS.CUSTOM_INVALID,
-      file => file,
-      result => result.valid,
-      result => result.message
-    ]
-  ].map(function ([validate, type, getValue, getValid, getMessage]) {
-    return function (file) {
-      return Promise.resolve(validate(file)).then(function (result) {
-        if (getValid(result)) {
-          return
-        }
-        return {
-          type,
-          value: getValue(file),
-          message: getMessage(result)
-        }
-      })
-    }
-  })
-
-  return function validateFile (file) {
-    return Promise.all(validators.map(validate => validate(file))).then(
-      function (errors) {
-        errors = errors.filter(identity)
-        return errors.length ? errors : null
-      }
-    )
-  }
-}
-
-class UploaderFile {
-  constructor (file) {
-    this.key = uniqueId('veuiUploaderFile')
-    this.status = UPLOAD_PENDING
-    this.message = undefined
-
-    this._file = file
-    this.meta = file
-      ? {
-        name: file.name,
-        type: getFileMediaType(file)
-      }
-      : {}
-
-    this.loaded = undefined
-    this.total = undefined
-  }
-
-  get isPending () {
-    return this.status === UPLOAD_PENDING
-  }
-
-  get isFailure () {
-    return this.status === UPLOAD_FAILURE
-  }
-
-  set isFailure (val) {
-    if (val) {
-      if (this.isUploading) {
-        this.cancel()
-      }
-      this.status = UPLOAD_FAILURE
-    }
-  }
-
-  get isUploading () {
-    return this.status === UPLOAD_UPLOADING
-  }
-
-  set isUploading (val) {
-    if (val) {
-      this.status = UPLOAD_UPLOADING
-    }
-  }
-
-  get isSuccess () {
-    return this.status === UPLOAD_SUCCESS
-  }
-
-  set isSuccess (val) {
-    if (val) {
-      if (this.isUploading) {
-        this.cancel()
-      }
-      this.status = UPLOAD_SUCCESS
-    }
-  }
-
-  get name () {
-    // TODO: eslint 该升级了！
-    // return this.result?.name ?? this.meta.name
-    return (this.result && this.result.name) || this.meta.name
-  }
-
-  get src () {
-    return (this.result && this.result.src) || this.meta.src
-  }
-
-  get poster () {
-    return (this.result && this.result.poster) || this.meta.poster
-  }
-
-  get type () {
-    return (this.result && this.result.type) || this.meta.type
-  }
-
-  set type (val) {
-    this.meta.type = val
-  }
-
-  get native () {
-    return this._file
-  }
-
-  get value () {
-    return {
-      _key: this.key,
-      ...pick(this, PUBLIC_FILE_PROPS),
-      ...this.extra
-    }
-  }
-
-  set value (val) {
-    this.isSuccess = true
-    if (val._key) {
-      this.key = val._key
-    }
-    this.result = pick(val, PUBLIC_FILE_PROPS)
-    this.extra = omit(val, ['_key', ...PUBLIC_FILE_PROPS])
-  }
-
-  static fromValue (val) {
-    let file = new UploaderFile()
-    file.value = val
-    return file
-  }
-
-  validate (options, context) {
-    const validateFile = getValidateFile(options, context)
-    return validateFile(this._file)
-  }
-
-  upload (context, callbacks) {
-    const listeners = ['onload', 'onerror', 'onprogress', 'oncancel'].reduce(
-      (ret, key) => {
-        ret[key] = (...args) =>
-          compact([this[key].bind(this), callbacks[key]]).forEach(execute =>
-            execute(...args)
-          )
-        return ret
-      },
-      {}
-    )
-    this._cancel = context.uploadRequest(this._file, listeners)
-  }
-
-  onload (data) {
-    this.status = data.success ? UPLOAD_SUCCESS : UPLOAD_FAILURE
-    this.result = omit(data, ['success'])
-
-    if (data.src) {
-      // 上传完成，不需要再 hold 文件对象
-      this._file = undefined
-    }
-  }
-
-  onerror (err) {
-    this.message = err.message
-    this.status = UPLOAD_FAILURE
-  }
-
-  onprogress (evt) {
-    this.loaded = evt.loaded
-    this.total = evt.total
-  }
-
-  oncancel () {}
-
-  cancel () {
-    if (this.status === UPLOAD_UPLOADING && this._cancel) {
-      this._cancel()
-      this._cancel = undefined
-    }
-  }
-}
-
-function getUploadRequest (options) {
-  const { requestMode, action, upload } = options
-  if (requestMode === 'xhr' && action) {
-    return getXhrUploadRequest(options)
-  } else if (requestMode === 'iframe' && action) {
-    return getIframeUploadRequest(options)
-  } else if (requestMode === 'custom' && upload) {
-    return getCustomUploadRequest(requestMode)
-  }
-  throw new Error('no matched upload method')
-}
-
-function getIframeUploadRequest (options) {
-  const parseResponse = getResonpseParse(options)
-  const { name, action, iframeMode, callbackNamespace, payload } = options
-
-  function getForm () {
-    const iframeId = uniqueId('veui-uploader-iframe')
-
-    let form = document.createElement('form')
-    form.method = 'POST'
-    form.action = action
-    form.target = iframeId
-    form.enctype = 'multipart/form-data'
-
-    let iframe = document.createElement('iframe')
-    iframe.name = iframeId
-    iframe.id = iframeId
-    iframe.hidden = true
-
-    document.body.appendChild(form)
-    document.body.appendChild(iframe)
-
-    function clean () {
-      document.body.removeChild(form)
-      document.body.removeChild(iframe)
-    }
-
-    return [form, iframe, clean]
-  }
-
-  function attachFileToForm (file, form) {
-    let fileInput = document.createElement('input')
-    fileInput.type = 'file'
-    fileInput.hidden = true
-    fileInput.name = name
-    // 解耦了视图和上传逻辑，但是提交到 iframe 需要把 pickFile 选到的文件塞回 input
-    // files 支持设置 FileList，但是 FileList 没有 slice 或者构造函数来从多文件 FileList 得到单文件 FileList
-    // 所以上面 pickFile 逻辑里保证了 iframe 情况下只能选单文件，并把 FileList 关联到 File 上
-    fileInput.files = file._rawFileList
-    form.appendChild(fileInput)
-  }
-
-  function attachObjectToForm (obj, form) {
-    entries(obj).forEach(function ([key, val]) {
-      let input = document.createElement('input')
-      input.type = 'hidden'
-      input.name = key
-      input.value = val
-      form.appendChild(input)
-    })
-  }
-
-  function bindPostMessageCallback (iframeId, ondata) {
-    function callback (event) {
-      if (
-        !event.source ||
-        !event.source.frameElement ||
-        event.source.frameElement.id !== iframeId
-      ) {
-        return
-      }
-      // 支持action为绝对路径或相对路径，ie9里的location没有origin
-      let actionOrigin = /^https?:\/\//.test(action)
-        ? action.match(/^https?:\/\/[^/]*/)[0]
-        : location.origin || location.protocol + '//' + location.host
-      if (actionOrigin === event.origin) {
-        ondata(event.data)
-      }
-    }
-    function cancel () {
-      window.removeEventListener('message', callback)
-    }
-    window.addEventListener('message', callback)
-    return cancel
-  }
-
-  function bindJsonpCallback (callbackFuncName, ondata) {
-    if (!window[callbackNamespace]) {
-      window[callbackNamespace] = {}
-    }
-    window[callbackNamespace][callbackFuncName] = ondata
-    return function cancel () {
-      window[callbackNamespace][callbackFuncName] = undefined
-    }
-  }
-
-  return function iframeUploadRequest (
-    file,
-    { onload, onerror, onprogress, oncancel }
-  ) {
-    const [form, iframe, cleanDom] = getForm()
-    attachFileToForm(file, form)
-    attachObjectToForm(payload, form)
-
-    let removeBind
-    if (iframeMode === 'postmessage') {
-      removeBind = bindPostMessageCallback(iframe.id, ondata)
-    } else if (iframeMode === 'callback') {
-      const callbackFuncName = uniqueId('veuiUploaderCallback')
-      removeBind = bindJsonpCallback(callbackFuncName, ondata)
-      attachObjectToForm(
-        { callback: `parent.${callbackNamespace}['${callbackFuncName}']` },
-        form
-      )
-    } else {
-      throw new Error('no matched iframe mode')
-    }
-
-    // TODO: timeout ?
-    function ondata (data) {
-      clean()
-      onprogress({
-        loaded: file.size,
-        total: file.size
-      })
-      onload(parseResponse(data))
-    }
-
-    function clean () {
-      removeBind()
-      cleanDom()
-    }
-
-    // TODO: fake progress?
-    onprogress({
-      loaded: 0,
-      total: file.size
-    }) // IE 的 ProgressEvent 不支持 constructor，所以这里只能抛个 plain object
-
-    form.submit()
-
-    return clean
-  }
-}
-
-function getXhrUploadRequest (options) {
-  const parseResponse = getResonpseParse(options)
-  const { name, action, headers, payload, withCredentials } = options
-  return function xhrUploadRequest (
-    file,
-    { onload, onerror, onprogress, oncancel }
-  ) {
-    let cancelled
-
-    let formData = new FormData()
-    formData.append(name, file)
-    entries(payload).forEach(function ([key, val]) {
-      formData.append(key, val)
-    })
-
-    let xhr = new XMLHttpRequest()
-    xhr.upload.onprogress = onprogress
-    xhr.onload = () => onload(parseResponse(xhr.responseText))
-    xhr.onerror = (...args) => !cancelled && onerror(...args)
-
-    xhr.open('POST', action, true)
-    entries(headers).forEach(function ([key, val]) {
-      xhr.setRequestHeader(key, val)
-    })
-    xhr.withCredentials = withCredentials
-    xhr.send(formData)
-
-    return function cancel () {
-      cancelled = true
-      xhr.abort()
-    }
-  }
-}
-
-function getCustomUploadRequest ({ upload }) {
-  // 之前的实现里自定义上传函数的 context 是 null，这里继续保留
-  return upload.bind(null)
-}
-
-function getResonpseParse ({ convertResponse = identity, dataType }) {
-  return function (data) {
-    if (typeof data === 'string' && dataType === 'json') {
-      try {
-        data = JSON.parse(data)
-      } catch (err) {
-        return convertResponse(undefined, err)
-      }
-    }
-    return convertResponse(data)
   }
 }
 </script>
